Fix dtype mismatch: cast loc_feature to float32

+++ src/model/NRFormer_Plus.py
@@ -636,7 +636,7 @@ class LocationEncoder(nn.Module):
         # Cache location tensor on device to avoid repeated tensor creation
         if self._loc_buffer is None or self._loc_buffer.device != next(self.parameters()).device:
             if isinstance(loc_feature, torch.Tensor):
-                self._loc_buffer = loc_feature.to(next(self.parameters()).device)
+                self._loc_buffer = loc_feature.float().to(next(self.parameters()).device)
             else:
                 self._loc_buffer = torch.tensor(loc_feature, dtype=torch.float32).to(next(self.parameters()).device)
         loc_embedding = self.loc_mlp(self._loc_buffer)
